remove iOS attempt for now
